refactor: split middleware configurations
