+++ src/middleware.ts
@@ -1,21 +1,10 @@
-import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
+import { auth } from "@/lib/middleware/auth.middleware";
 
 export function middleware(req: NextRequest) {
-  const sessionId = req.cookies.get("sessionId")?.value;
-  const { pathname } = req.nextUrl;
-
-  if (!sessionId && pathname.startsWith("/dashboard")) {
-    return NextResponse.redirect(new URL("/login", req.url));
-  }
-
-  if (sessionId && pathname.startsWith("/login")) {
-    return NextResponse.redirect(new URL("/dashboard", req.url));
-  }
-
-  return NextResponse.next();
+  return auth.middleware(req);
 }
 
 export const config = {
-  matcher: ["/login", "/dashboard/:path*"],
-};
+  matcher: [...auth.matcher],
+} as const;
